Restructure model definition for improved flexibility
