+++ src/main/scala/com/dataintuitive/luciuscore/io/RanksIO.scala
@@ -1,6 +1,6 @@
 package com.dataintuitive.luciuscore.io
 
-import com.dataintuitive.luciuscore.Model._
+import com.dataintuitive.luciuscore.model.v3.Model._
 import com.dataintuitive.luciuscore.TransformationFunctions._
 import org.apache.spark.rdd.RDD
