+++ src/main/scala/com/dataintuitive/luciuscore/io/StatsIO.scala
@@ -1,7 +1,7 @@
 package com.dataintuitive.luciuscore.io
 
 import IoFunctions._
-import com.dataintuitive.luciuscore.Model.DbRow
+import com.dataintuitive.luciuscore.model.v3.Model.DbRow
 import org.apache.spark.SparkContext
 import com.dataintuitive.luciuscore.utilities.RddFunctions._
 import org.apache.spark.rdd.RDD
