+++ src/main/scala/com/dataintuitive/luciuscore/lenses/DbRowLenses.scala
@@ -1,8 +1,8 @@
 package com.dataintuitive.luciuscore.lenses
 
 import scalaz.Lens
-import com.dataintuitive.luciuscore.Model._
-import com.dataintuitive.luciuscore.Filter
+import com.dataintuitive.luciuscore.model.v3.Model._
+import com.dataintuitive.luciuscore.model.v3.Filter
 
 object DbRowLenses extends Serializable {
