+++ src/main/scala/com/dataintuitive/luciuscore/lenses/SampleLenses.scala
@@ -1,7 +1,7 @@
 package com.dataintuitive.luciuscore.lenses
 
 import scalaz.Lens
-import com.dataintuitive.luciuscore.Model._
+import com.dataintuitive.luciuscore.model.v3.Model._
 
 object SampleLenses extends Serializable {
