+++ src/main/scala/com/dataintuitive/luciuscore/package.scala
@@ -5,7 +5,7 @@ package com.dataintuitive
   */
 package object luciuscore {
 
-  import com.dataintuitive.luciuscore.Model._
+  import com.dataintuitive.luciuscore.model.v3.Model._
   import com.dataintuitive.luciuscore.Implicits._
   import com.dataintuitive.luciuscore.io._
